requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lance-namespace==0.5.2
lance-namespace-urllib3-client==0.5.2
numpy==2.4.2
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pylance==2.0.1
pyTelegramBotAPI==4.31.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
requests==2.32.5
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
